init refactor.
